requirements.txt
# --- Servidor Web ---
Flask

# --- APIs de Google ---
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib


# --- Manejo de Excel ---
openpyxl

# --- Manejo de Imágenes ---
Pillow

# --- Variables de Entorno Locales ---
python-dotenv